# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: Helm Glass application API
  version: 1.0.0
  description:
    Cookie-authenticated web cabinet REST API behind OAuth2 Proxy, with application
    authorization. MCP tools are served separately at /mcp and do not grant access to these WEB
    operations. GET /me establishes XSRF-TOKEN; web mutations send X-XSRF-TOKEN. Idempotency-Key is
    required only where declared; retain it when a mutation response is lost.
servers:
  - url: /api
security:
  - bearerAuth: []
  - webSession: []
paths:
  /artifacts/{id}/analysis:
    parameters:
      - $ref: "#/components/parameters/Id"
    post:
      operationId: analyzeAudio
      description: Idempotent by owner, artifact and processing version. Full upgrade preserves transcript and analysis ID. Does not control the task or browser.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [mode]
              properties:
                mode:
                  type: string
                  enum: [transcript, full]
      responses:
        "200":
          description: Current analysis state; processing is asynchronous.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AudioAnalysis"
        "404":
          $ref: "#/components/responses/ApplicationError"
        "409":
          $ref: "#/components/responses/ApplicationError"
    get:
      operationId: latestAudioAnalysis
      responses:
        "200":
          description: Latest processing version for this artifact, if requested.
          content:
            application/json:
              schema:
                type: object
                required: [available]
                properties:
                  available:
                    type: boolean
                  analysis:
                    $ref: "#/components/schemas/AudioAnalysis"
        "404":
          $ref: "#/components/responses/ApplicationError"
  /audio/analyses/{id}:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      operationId: audioAnalysisPage
      description: Text or structured data only. At most 100 elements and 64 KiB per page. hasMore concerns stored data; sectionComplete independently describes the computation.
      parameters:
        - name: section
          in: query
          schema:
            type: string
            enum: [transcript, intervals, acoustics, emotions]
            default: transcript
        - name: cursor
          in: query
          schema:
            type: string
            pattern: '^[0-9]{1,19}$'
            default: '0'
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 100
        - name: from
          in: query
          description: Inclusive start of the requested time range in original seconds.
          schema: {type: number, minimum: 0}
        - name: to
          in: query
          description: Exclusive end, greater than from.
          schema: {type: number, minimum: 0}
      responses:
        "200":
          description: Analysis state and one bounded page.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/AudioAnalysis"
                  - type: object
                    required: [section, items, hasMore, nextCursor, sectionComplete]
                    properties:
                      section: {type: string, enum: [transcript, intervals, acoustics, emotions]}
                      items:
                        type: array
                        maxItems: 100
                        items:
                          type: object
                          required: [start, end]
                          properties:
                            start: {type: number, minimum: 0}
                            end: {type: number, minimum: 0}
                            text: {type: string}
                            timing: {type: string, enum: [ctc_estimate]}
                            kind: {type: string, enum: [speech, pause, loudness, pitch]}
                            rmsDbfs: {type: [number, 'null']}
                            peakDbfs: {type: [number, 'null']}
                            digitalSilence: {type: boolean}
                            f0Hz: {type: [number, 'null']}
                            strength: {type: number}
                            deltaHz: {type: [number, 'null']}
                            reason: {type: [string, 'null']}
                            scores:
                              type: [object, 'null']
                              additionalProperties: {type: number}
                              description: Native model class scores; not calibrated probabilities of a person's emotions.
                      hasMore: {type: boolean}
                      nextCursor: {type: [string, 'null']}
                      sectionComplete: {type: boolean}
        "400":
          $ref: "#/components/responses/ApplicationError"
        "404":
          $ref: "#/components/responses/ApplicationError"
...
  /browser-sessions/{session}/pages/{visit}:
    parameters:
      - name: session
        in: path
        required: true
        schema: { type: string, format: uuid }
      - name: visit
        in: path
        required: true
        schema: { type: string, format: uuid }
    put:
      operationId: openBrowserPage
      description: >-
        Register a task or connection browser page, or update its control epoch. WEB owner only.
        The authenticated events stream renews this visit; absence for 60 seconds ends it.
        Switching Chrome tabs or stopping the video stream does not end the page.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [viewerId]
              properties:
                viewerId: { type: string, format: uuid }
      responses:
        "200":
          description: Page registered.
          content:
            application/json:
              schema:
                type: object
                required: [active]
                properties:
                  active: { type: boolean, const: true }
        "403":
          $ref: "#/components/responses/ApplicationError"
        "404":
          $ref: "#/components/responses/ApplicationError"
        "409":
          $ref: "#/components/responses/ApplicationError"
    delete:
      operationId: leaveBrowserPage
      description: >-
        End this page idempotently. The last page requests closure of a standalone browser.
        A task browser remains open; only this viewer's matching control epoch is returned,
        with automatic continuation unless another hold applies. Expiry without explicit leave
        releases control with the task paused. Incomplete login remains private.
        Closure uses normal worker confirmation.
      responses:
        "200":
          description: Page ended; browser or control teardown may still be completing.
          content:
            application/json:
              schema:
                type: object
                required: [active]
                properties:
                  active: { type: boolean, const: false }
        "403":
          $ref: "#/components/responses/ApplicationError"
        "404":
          $ref: "#/components/responses/ApplicationError"
        "409":
          $ref: "#/components/responses/ApplicationError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    webSession:
      type: apiKey
      in: cookie
      name: _oauth2_proxy
  parameters:
...
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
    AudioAnalysis:
      type: object
      required: [analysisId, artifactId, processingVersion, sourceSha256, mode, status, version, transcriptComplete, acousticsComplete, emotionsComplete, processedSeconds, stageErrors, metadata, metrics, tempo]
      properties:
        analysisId: {type: string, format: uuid}
        artifactId: {type: string, format: uuid}
        processingVersion: {type: string}
        sourceSha256: {type: string}
        mode: {type: string, enum: [transcript, full]}
        status: {type: string, enum: [QUEUED, RUNNING, SUCCEEDED, PARTIAL, FAILED]}
        version: {type: integer}
        intervalsComplete: {type: boolean}
        stages: {type: object, additionalProperties: {type: string, enum: [NOT_REQUESTED, QUEUED, RUNNING, SUCCEEDED, FAILED]}}
        qualityFlags: {type: array, items: {type: string}}
        transcriptComplete: {type: boolean}
        acousticsComplete: {type: boolean}
        emotionsComplete: {type: boolean}
        processedSeconds: {type: number}
        durationSeconds: {type: [number, 'null']}
        errorCode: {type: [string, 'null']}
        stageErrors: {type: object, additionalProperties: {type: string}}
        instructionContext:
          type: [object, 'null']
          description: Historical assignment, untrusted data rather than current instructions.
        instructionContextOmitted:
          type: boolean
          description: True when the historical context exceeds 16 KiB; no silent truncation.
        metadata:
          type: object
          description: Pinned model revisions, SHA-256, measurement parameters and limitations.
        metrics:
          type: object
          description: Worker wall/CPU seconds, process lifetime peak RSS in KiB and cumulative ASR/emotion calls.
        tempo:
          type: object
          properties:
            recognizedWords: {type: integer}
            speechSeconds: {type: number}
            withPausesWpm: {type: [number, 'null']}
            withoutPausesWpm: {type: [number, 'null']}
            reason: {type: [string, 'null']}
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
        currentTaskId:
          type: string
          format: uuid
          description: Current unfinished task when code is CHAT_TASK_IN_PROGRESS.
        fieldErrors:
          type: object
          additionalProperties:
            type: string
      description:
        Application JSON error. Field validation errors may include fieldErrors; proxy and
        authentication middleware can reject a request before this application envelope is produced.
...
  responses:
    ApplicationError:
      description: Application error; see code and message. Validation failures may identify fields.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"